# contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Atomic Bank Operations API
  description: Production-safe banking microservice with atomic operations, auditable ledger, and event-driven architecture
  version: 1.0.0
  contact:
    name: Bank Operations Team
    email: ops@bank.example.com
  license:
    name: Proprietary
    url: https://bank.example.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.bank.example.com
    description: Production server
paths:
...
  /v1/audit/transactions:
    get:
      summary: Get all transactions (operator access)
      operationId: getAllTransactions
      tags:
        - Audit
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '403':
          description: Insufficient permissions (requires OPERATOR or ADMIN)
  /v1/audit/ledger/{accountId}:
    get:
      summary: Get ledger entries for account
      operationId: getLedgerByAccount
      tags:
        - Audit
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          description: Account UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerEntryResponse'
        '403':
          description: Insufficient permissions
  /v1/audit/ledger:
    get:
      summary: Get complete ledger (admin access only)
      operationId: getCompleteLedger
      tags:
        - Audit
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Complete ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerEntryResponse'
        '403':
          description: Insufficient permissions (requires ADMIN)
  /v1/audit/ledger/transaction/{transactionId}:
    get:
      summary: Get ledger entries for transaction
      operationId: getLedgerByTransaction
      tags:
        - Audit
      security:
        - BearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          description: Transaction UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction ledger summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLedgerSummary'
        '404':
          description: Transaction not found
  /v1/audit/statistics:
    get:
      summary: Get transaction statistics
      operationId: getStatistics
      tags:
        - Audit
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Transaction statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
components:
  schemas:
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Transaction ID
        type:
          type: string
          enum: [WITHDRAWAL, TRANSFER, DEPOSIT, FEE]
          description: Transaction type
        sourceAccountId:
          type: string
          format: uuid
          description: Source account ID (nullable for deposit)
        destinationAccountId:
          type: string
          format: uuid
          description: Destination account ID (nullable for withdrawal)
        status:
          type: string
          enum: [PENDING, POSTED, FAILED]
          description: Transaction status
        amount:
          type: number
          format: decimal
          description: Transaction amount
        currency:
          type: string
          description: Currency code
        idempotencyKey:
          type: string
          format: uuid
          description: Idempotency key
        initiatedBy:
          type: string
          description: User or service that initiated transaction
        createdAt:
          type: string
          format: date-time
          description: Transaction creation timestamp
        completedAt:
          type: string
          format: date-time
          description: Transaction completion timestamp (nullable if pending)
    LedgerEntryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Ledger entry ID
        transactionId:
          type: string
          format: uuid
          description: Associated transaction ID
        accountId:
          type: string
          format: uuid
          description: Associated account ID
        type:
          type: string
          enum: [DEBIT, CREDIT, FEE]
          description: Entry type
        amount:
          type: number
          format: decimal
          description: Entry amount
        currency:
          type: string
          description: Currency code
        timestamp:
          type: string
          format: date-time
          description: Entry timestamp
    TransactionLedgerSummary:
      type: object
      properties:
        transaction:
          $ref: '#/components/schemas/TransactionResponse'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntryResponse'
        totalDebits:
          type: number
          format: decimal
          description: Sum of debit entries
        totalCredits:
          type: number
          format: decimal
          description: Sum of credit entries
        totalFees:
          type: number
          format: decimal
          description: Sum of fee entries
        isBalanced:
          type: boolean
          description: Whether transaction maintains double-entry bookkeeping
    StatisticsResponse:
      type: object
      properties:
        totalTransactions:
          type: integer
          description: Total number of transactions
        postedTransactions:
          type: integer
          description: Number of successfully posted transactions
        pendingTransactions:
          type: integer
          description: Number of pending transactions
        failedTransactions:
          type: integer
          description: Number of failed transactions
        totalPostedAmount:
          type: number
          format: decimal
          description: Sum of all posted transaction amounts
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /v1/auth/login
...